+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ## Changed
 
+- Added `:kaocha.report/printed-expression` to the `debug` reporter, for
+  debugging reporting issues with kaocha-cljs
+
 # 1.0-612 (2020-03-29 / 06293c8)
 
 ## Added